moved unified_model driver specific code in module and add unit tests
