+++ forest/data.py
@@ -1,14 +1,8 @@
-import os
-import datetime as dt
-import re
 try:
     import cartopy
 except ImportError:
     # ReadTheDocs unable to pip install cartopy
     pass
-import glob
-import json
-import pandas as pd
 import numpy as np
 import netCDF4
 try:
@@ -16,26 +10,20 @@ try:
 except ImportError:
     # ReadTheDocs unable to pip install cf-units
     pass
-from forest import (
-        geo,
-        disk)
+from forest import geo
 import bokeh.models
 from collections import OrderedDict
 from functools import partial
-import scipy.ndimage
 try:
     import shapely.geometry
 except ImportError:
     # ReadTheDocs unable to pip install shapely
     pass
 from forest.util import (
-        initial_time,
-        coarsify)
-from forest.exceptions import SearchFail
+        initial_time)
 
 
 # Application data shared across documents
-IMAGES = OrderedDict()
 VECTORS = OrderedDict()
 COASTLINES = {
     "xs": [],
@@ -126,6 +114,7 @@ def iterlines(geometries):
             yield xy(geometry)
 
 
+# TODO: Delete this decorator in a future PR
 def cache(name):
     store = globals()[name]
     def decorator(f):
@@ -136,7 +125,7 @@ def cache(name):
         return wrapped
     return decorator
 
-
+# TODO: Delete this class in a future PR
 class ActiveViewer(object):
     def __init__(self):
         self.active = False
@@ -167,6 +156,7 @@ class ActiveViewer(object):
             self.pending_state = state
 
 
+# TODO: Delete this class in a future PR
 class WindBarbs(ActiveViewer):
     def __init__(self, paths):
         self.paths = paths
@@ -246,6 +236,7 @@ class WindBarbs(ActiveViewer):
         }
 
 
+# TODO: Delete this class in a future PR
 class Finder(object):
     def __init__(self, paths):
         self.paths = paths
@@ -295,119 +286,6 @@ class Finder(object):
                     np.abs(
                         self.initial_times - initial_time))
             return self.paths[i]
-
-
-def pts_hash(pts):
-    if isinstance(pts, np.ndarray):
-        return pts.tostring()
-    else:
-        return pts
-
-
-def load_image(path, variable, itime, ipressure):
-    return load_image_pts(path, variable, (itime,), (itime, ipressure))
-
-
-def _image_cache(hash_func):
-    """Simple wrapper to cache images per-server"""
-    def inner(load_func):
-        def innermost(*args):
-            key = hash_func(*args)
-            if key not in IMAGES:
-                # TODO: Replace this infinite cache
-                IMAGES[key] = load_func(*args)
-            total_bytes = 0
-            for data in IMAGES.values():
-                total_bytes +=  _image_size(data)
-            print("Total bytes", total_bytes)
-            return IMAGES[key]
-        return innermost
-    return inner
-
-
-def _image_hash(path, variable, pts_3d, pts_4d):
-    """Convert arguments to hashable type"""
-    return (path, variable, pts_hash(pts_3d), pts_hash(pts_4d))
-
-
-def _image_size(data):
-    # Approx. image payload size
-    return sum(arr.nbytes for arr in data["image"])
-
-
-@_image_cache(_image_hash)
-def load_image_pts(path, variable, pts_3d, pts_4d):
-    """Load bokeh image glyph data from file using slices"""
-    try:
-        lons, lats, values, units = _load_netcdf4(path, variable, pts_3d, pts_4d)
-    except:
-        lons, lats, values, units = _load_cube(path, variable, pts_3d, pts_4d)
-
-    # Units
-    if variable in ["precipitation_flux", "stratiform_rainfall_rate"]:
-        if units == "mm h-1":
-            values = values
-        else:
-            values = convert_units(values, units, "kg m-2 hour-1")
-    elif units == "K":
-        values = convert_units(values, "K", "Celsius")
-
-    # Coarsify images
-    threshold = 200 * 200  # Chosen since TMA WRF is 199 x 199
-    if values.size > threshold:
-        fraction = 0.25
-    else:
-        fraction = 1.
-    lons, lats, values = coarsify(
-        lons, lats, values, fraction)
-
-    # Roll input data into [-180, 180] range
-    if np.any(lons > 180.0):
-        shift_by = np.sum(lons > 180.0)
-        lons[lons > 180.0] -= 360.
-        lons = np.roll(lons, shift_by)
-        values = np.roll(values, shift_by, axis=1)
-
-    return geo.stretch_image(lons, lats, values)
-
-
-def _load_cube(path, variable, pts_3d, pts_4d):
-    import iris
-    cube = iris.load_cube(path, iris.Constraint(variable))
-    units = cube.units
-    lons = cube.coord('longitude').points
-    if lons.ndim == 2:
-        lons = lons[0, :]
-    lats = cube.coord('latitude').points
-    if lons.ndim == 2:
-        lats = lats[:, 0]
-    if cube.data.ndim == 4:
-        values = cube.data[pts_4d]
-    else:
-        values = cube.data[pts_3d]
-    return lons, lats, values, units
-
-
-def _load_netcdf4(path, variable, pts_3d, pts_4d):
-    with netCDF4.Dataset(path) as dataset:
-        try:
-            var = dataset.variables[variable]
-        except KeyError as e:
-            if variable == "precipitation_flux":
-                var = dataset.variables["stratiform_rainfall_rate"]
-            else:
-                raise e
-        for d in var.dimensions:
-            if "longitude" in d:
-                lons = dataset.variables[d][:]
-            if "latitude" in d:
-                lats = dataset.variables[d][:]
-        if len(var.dimensions) == 4:
-            values = var[pts_4d]
-        else:
-            values = var[pts_3d]
-        units = var.units
-    return lons, lats, values, units
 
 
 def convert_units(values, old_unit, new_unit):
