+++ test/test_drivers_unified_model.py
@@ -1,7 +1,11 @@
+import pytest
 import bokeh.models
 import forest.drivers
+from forest.drivers import unified_model
 import forest.db
 import sqlite3
+import netCDF4
+import iris
 
 
 def test_dataset_loader_pattern():
@@ -44,3 +48,37 @@ def test_loader_use_database(tmpdir):
     view = dataset.map_view()
     assert hasattr(view.loader.locator, "connection")
     assert view.loader.locator.directory == "/replace"
+
+
+def test_load_image_pts(tmpdir):
+    path = str(tmpdir / "file.nc")
+    variable = "air_temperature"
+    with netCDF4.Dataset(path, "w") as dataset:
+        make_file(dataset, variable)
+    data = unified_model.load_image_pts(path, variable, (), ())
+    assert data["image"][0].shape == (2, 2)
+
+
+def test_iris_load(tmpdir):
+    path = str(tmpdir / "file.nc")
+    variable = "air_temperature"
+    with netCDF4.Dataset(path, "w") as dataset:
+        make_file(dataset, variable)
+    cubes = iris.load(path)
+    names = [c.name() for c in cubes]
+    assert names == [variable]
+
+
+def make_file(dataset, variable):
+    dimensions = [
+        ("time", 1),
+        ("longitude", 2),
+        ("latitude", 2)
+    ]
+    for name, length in dimensions:
+        dataset.createDimension(name, length)
+    var = dataset.createVariable("longitude", "f", ("longitude",))
+    var[:] = [0, 1]
+    var = dataset.createVariable("latitude", "f", ("latitude",))
+    var[:] = [0, 1]
+    var = dataset.createVariable(variable, "f", ("longitude", "latitude"))
